packages/shared-auth-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "raysourcelabs-shared-auth"
version = "1.0.0"
description = "Shared authentication utilities for Raysource Labs applications"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Raysource Labs", email = "dev@raysourcelabs.com" }
]
requires-python = ">=3.10"
dependencies = [
    "pyjwt>=2.8.0",
    "bcrypt>=4.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.hatch.build.targets.wheel]
packages = ["shared_auth"]
